stability mods + ENUMS class
plugin now load on webview's startup not on first execution.
MFCCHandlerThread is started on bounding, in the MFCCService::initService()

AudioInputReceiver receives a second new handler, which send data directly to MFCCHandlerThread.
so we ease the data transmission.
this step was critical as SOMETIMES the internal handler of MFCCHandlerThread was null.
that's why, it is started after service bounding and the plugin is initialized on App start. this seems to work....monitor it !!!!

complete cleanup of between threads messages' action_codes. now there is an ENUMS class containing all the unique codes.
JS interface must adhere to it.